requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
brotli==1.0.7             # via flask-compress
click==7.1.2              # via flask
dash-core-components==1.9.1  # via dash
dash-html-components==1.0.3  # via dash
dash-table==4.6.2         # via dash
dash==1.11.0              # via -r requirements.in
dash_renderer==1.4.0      # via dash
flask-compress==1.5.0     # via dash
flask==1.1.2              # via dash, flask-compress
future==0.18.2            # via dash
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
numpy==1.18.4             # via pandas
pandas==1.0.3             # via -r requirements.in
plotly==4.6.0             # via dash
psycopg2-binary==2.8.5    # via -r requirements.in
python-dateutil==2.8.1    # via pandas
python-dotenv==0.13.0     # via -r requirements.in
pytz==2020.1              # via pandas
retrying==1.3.3           # via plotly
six==1.14.0               # via plotly, python-dateutil, retrying
sqlalchemy==1.3.15        # via -r requirements.in
werkzeug==0.16.0          # via -r requirements.in, flask
